+++ lerobot/common/robot_devices/robots/utils.py
@@ -5,7 +5,7 @@ from lerobot.common.robot_devices.robots.configs import (
     KochBimanualRobotConfig,
     KochRobotConfig,
     ManipulatorRobotConfig,
-    MobileSO100RobotConfig,
+    LeKiwiRobotConfig,
     MossRobotConfig,
     RobotConfig,
     So100RobotConfig,
@@ -46,8 +46,8 @@ def make_robot_config(robot_type: str, **kwargs) -> RobotConfig:
         return So100RobotConfig(**kwargs)
     elif robot_type == "stretch":
         return StretchRobotConfig(**kwargs)
-    elif robot_type == "mobile_so100":
-        return MobileSO100RobotConfig(**kwargs)
+    elif robot_type == "lekiwi":
+        return LeKiwiRobotConfig(**kwargs)
     else:
         raise ValueError(f"Robot type '{robot_type}' is not available.")
 
@@ -57,7 +57,7 @@ def make_robot_from_config(config: RobotConfig):
         from lerobot.common.robot_devices.robots.manipulator import ManipulatorRobot
 
         return ManipulatorRobot(config)
-    elif isinstance(config, MobileSO100RobotConfig):
+    elif isinstance(config, LeKiwiRobotConfig):
         from lerobot.common.robot_devices.robots.mobile_manipulator import MobileManipulator
 
         return MobileManipulator(config)
