+++ lerobot/scripts/control_robot.py
@@ -77,6 +77,13 @@ python lerobot/scripts/control_robot.py record \
     --control.reset_time_s=10
 ```
 
+- Run the code for the LeKiwi robot (on the robot itself),
+```bash
+python lerobot/scripts/control_robot.py \
+  --robot.type=lekiwi \
+  --control.type=remote_robot
+```
+
 **NOTE**: You can use your keyboard to control data recording flow.
 - Tap right arrow key '->' to early exit while recording an episode and go to resseting the environment.
 - Tap right arrow key '->' to early exit while resetting the environment and got to recording the next episode.
@@ -127,6 +134,7 @@ from lerobot.common.robot_devices.control_configs import (
     CalibrateControlConfig,
     ControlPipelineConfig,
     RecordControlConfig,
+    RemoteRobotConfig,
     ReplayControlConfig,
     TeleoperateControlConfig,
 )
@@ -188,13 +196,13 @@ def calibrate(robot: Robot, cfg: CalibrateControlConfig):
     if robot.is_connected:
         robot.disconnect()
 
-    if robot.robot_type.startswith("mobile_so100") and "main_follower" in arms:
-        print("Calibrating only the mobile_so100 follower arm 'main_follower'...")
+    if robot.robot_type.startswith("lekiwi") and "main_follower" in arms:
+        print("Calibrating only the lekiwi follower arm 'main_follower'...")
         robot.calibrate_follower()
         return
 
-    if robot.robot_type.startswith("mobile_so100") and "main_leader" in arms:
-        print("Calibrating only the mobile_so100 leader arm 'main_leader'...")
+    if robot.robot_type.startswith("lekiwi") and "main_leader" in arms:
+        print("Calibrating only the lekiwi leader arm 'main_leader'...")
         robot.calibrate_leader()
         return
 
@@ -367,6 +375,9 @@ def control_robot(cfg: ControlPipelineConfig):
         record(robot, cfg.control)
     elif isinstance(cfg.control, ReplayControlConfig):
         replay(robot, cfg.control)
+    elif isinstance(cfg.control, RemoteRobotConfig):
+        from lerobot.common.robot_devices.robots.lekiwi_remote import run_lekiwi
+        run_lekiwi(cfg.robot)
 
     if robot.is_connected:
         # Disconnect manually to avoid a "Core dump" during process
