+++ pyproject.toml
@@ -38,7 +38,7 @@ dependencies = [
     "einops>=0.8.0",
     "flask>=3.0.3",
     "gdown>=5.1.0",
-    "gymnasium==0.29.1",  # TODO(rcadene, aliberts): Make gym 1.0.0 work
+    "gymnasium==0.29.1", # TODO(rcadene, aliberts): Make gym 1.0.0 work
     "h5py>=3.10.0",
     "huggingface-hub[hf-transfer,cli]>=0.27.1 ; python_version < '4.0'",
     "hydra-core>=1.3.2",
@@ -49,12 +49,14 @@ dependencies = [
     "opencv-python>=4.9.0",
     "pyav>=12.0.5",
     "pymunk>=6.6.0",
+    "pynput>=1.7.7",
+    "pyzmq>=26.2.1",
     "rerun-sdk>=0.21.0",
     "termcolor>=2.4.0",
     "torch>=2.2.1",
     "torchvision>=0.21.0",
     "wandb>=0.16.3",
-    "zarr>=2.17.0"
+    "zarr>=2.17.0",
 ]
 
 [project.optional-dependencies]
